refactor: remove unused OrderInfo API and update imports in PendingOrders
